cambios a WORDMAP y eficiencia

+++ Kernel/include/mman.h
@@ -6,7 +6,7 @@
 
 void mm_init();
 
-void * my_malloc(uint64_t size);
+void * my_malloc(uint16_t size);
 void my_free(void * addr_to_free);
 
 uint64_t get_mem_total();
